+++ kodecks/src/player.rs
@@ -273,7 +273,7 @@ impl Score for Player {
     fn score(&self) -> PlayerScore {
         PlayerScore {
             id: self.id,
-            life: self.stats.life as i32,
+            life: self.stats.life,
             shards: self.shards.len() as i32,
             hand: self.hand.items().map(Score::score).collect(),
             field: self.field.items().map(Score::score).collect(),
@@ -283,7 +283,7 @@ impl Score for Player {
 
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, Encode, Decode)]
 pub struct PlayerStats {
-    pub life: u32,
+    pub life: i32,
 }
 
 impl Default for PlayerStats {
